Refactor runtime helpers and update documentation
