+++ cad_quoter/app/__init__.py
@@ -1,8 +1,11 @@
 """Application-level helpers for the CAD quoting tool."""
+from __future__ import annotations
 
 from .container import ServiceContainer, create_default_container
+from . import runtime
 
 __all__ = [
     "ServiceContainer",
     "create_default_container",
+    "runtime",
 ]
